+++ CHANGES.rst
@@ -5,8 +5,6 @@ Changelog
 1.3.4 (unreleased)
 ------------------
 
-- Adapted code to receive the smallest JSON possible by using
-  include parameters in the json query.
 - Query categories from iA.Delib to populate vocabularies only when loading the edit form.
   [odelaere]
 - Added automatic initialization of categories mapping.
@@ -17,6 +15,9 @@ Changelog
   [gbastien]
 - Added default value for `Institution.meeting_config_id`
   [odelaere]
+- Adapted code to receive the smallest JSON possible by using
+  include parameters in the json query.
+  [gbastien]
 
 
 1.3.3 (2021-06-28)